LazyQuery.jl

This package is currently a proof-of-concept for lazy-evaluation based table querying. It is built on LazyContext.jl. It only supports DataFrames at the moment, but there is no reason the package can't be extended to other tabular data structures.

julia>  import LazyContext, ChainRecursive, LazyQuery, DataFrames;

julia>  LazyContext.@new_environment;

julia>  LazyContext.@use_in_environment DataFrames LazyQuery;

julia>  ChainRecursive.@chain LazyContext.@evaluate begin
            DataFrame(
                a = [1, 1, 2, 2, 3, 3],
                be = [1, 3, 2, 4, 3, 5],
                C = [5, 3, 4, 2, 3, 1] )
            @rename it b = be c = C
            @add_to(it,
                d = b .+ c,
                e = b .- a)
            @choose_from it -c
            @rows_where it a .> 1
            @groupby it e
            map(it) do subframe
                @make_from(subframe,
                    f = sum(a),
                    g = mean(b)
                )
            end
            combine(it)
        end
2×3 DataFrames.DataFrame
│ Row │ e │ f │ g   │
├─────┼───┼───┼─────┤
│ 1   │ 0 │ 5 │ 2.5 │
│ 2   │ 2 │ 5 │ 4.5 │

Note that the anonymous do block above is slow; a new environment is set up containing the contents of a subframe is set up each time it is called. In fact, given a bit more work, @make_from could be called directly on a GroupedDataFrame. At the moment, this package is in the proof-of-concept stage.

LazyQuery.add_to

LazyQuery.choose_from

LazyQuery.make_from

LazyQuery.rows_where

LazyQuery.with

LazyQuery.add_to — Method.

add_to(w::LazyContext.WithContext, args...)

Add to w, evaluating args in context.

julia> import LazyContext, LazyQuery

julia> LazyContext.@new_environment;

julia> LazyContext.@use_in_environment LazyQuery;

julia> LazyContext.@evaluate begin
           d = Dict(:a => 1, :b => 2);
           @add_to d c = a + b
       end == Dict(:a => 1, :b => 2, :c => 3)
true

source

LazyQuery.choose_from — Method.

choose_from(w::LazyContext.WithContext, args...)

Evaluate with_context in an environment including w.

julia> import LazyContext, LazyQuery, DataFrames

julia> LazyContext.@new_environment;

julia> LazyContext.@use_in_environment DataFrames LazyQuery;

julia> LazyContext.@evaluate begin
           d = DataFrame(a = 1, b = 2, c = 3, d = 4)
           (@choose_from d a c:d) == (@choose_from d -b)
       end
true

source

LazyQuery.make_from — Method.

make_from(w::LazyContext.WithContext, args...)

Make a new object from w, evaluating args in context.

julia> import LazyContext, LazyQuery

julia> LazyContext.@new_environment;

julia> LazyContext.@use_in_environment LazyQuery;

julia> LazyContext.@evaluate begin
           d = Dict(:a => 1, :b => 2)
           @make_from(d, c = a + b, d = b - a)
       end == Dict(:c => 3, :d => 1)
true

source

LazyQuery.rows_where — Method.

rows_where(w::LazyContext.WithContext, rows)

Get rows from w, evaluating rows in context.

julia> import LazyContext, LazyQuery, DataFrames

julia> LazyContext.@new_environment;

julia> LazyContext.@use_in_environment LazyQuery DataFrames;

julia> LazyContext.@evaluate begin
           d = DataFrame(a = [1, 2, 3], b = [3, 2, 1] )
           @rows_where d a .== b
       end
1×2 DataFrames.DataFrame
│ Row │ a │ b │
├─────┼───┼───┤
│ 1   │ 2 │ 2 │

source

LazyQuery.with — Method.

with(w::LazyContext.WithContext, with_context)

Evaluate with_context in an environment including w.

julia> import LazyContext, LazyQuery

julia> LazyContext.@new_environment;

julia> LazyContext.@use_in_environment LazyQuery;

julia> LazyContext.@evaluate begin
           d = Dict(:a => 1, :b => 2)
           @with d a + b
       end
3

All other lazy functions in the package are based on with.

source